# src/adapter/driver/http/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Video API",
        "version": "1.0.0"
    },
    "security": [
        { "bearerAuth": [] }
    ],
    "paths": {
        "/video/upload": {
            "post": {
                "summary": "Upload de vídeo",
                "description": "Faz upload de um arquivo de vídeo e retorna os dados do vídeo salvo.",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "video": {
                                        "type": "string",
                                        "format": "binary",
                                        "description": "Arquivo de vídeo a ser enviado."
                                    }
                                },
                                "required": [
                                    "video"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Upload assíncrono recebido.",
                        "headers": {
                            "Location": {
                                "description": "Rota para consultar o status do vídeo enviado.",
                                "schema": {
                                    "type": "string",
                                    "format": "uri"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AsyncUploadPresenter"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Nenhum arquivo enviado ou arquivo inválido.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "examples": {
                                    "NoFile": {
                                        "summary": "Nenhum arquivo enviado",
                                        "value": { "statusCode": 400, "message": "No video file uploaded" }
                                    },
                                    "InvalidFile": {
                                        "summary": "Arquivo inválido",
                                        "value": { "statusCode": 400, "message": "File is empty or does not exist." }
                                    }
                                }
                            }
                        }
                    },
                    "413": {
                        "description": "Arquivo muito grande.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "examples": {
                                    "FileTooLarge": {
                                        "summary": "Arquivo excede o limite de tamanho",
                                        "value": { "statusCode": 413, "message": "Tamanho de arquivo excedido" }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "examples": {
                                    "SQS": {
                                        "summary": "Erro ao enviar para SQS",
                                        "value": { "statusCode": 500, "message": "Erro seguir para etapa de processamento do video" }
                                    },
                                    "Generic": {
                                        "summary": "Erro interno genérico",
                                        "value": { "statusCode": 500, "message": "Erro ao processar arquivo" }
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Video"
                ]
            }
        },
        "/video": {
            "get": {
                "summary": "Lista todos os vídeos",
                "description": "Retorna uma lista de vídeos do cliente autenticado.",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Lista de vídeos.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": { "$ref": "#/components/schemas/VideoPresenter" }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nenhum vídeo encontrado.",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                                "examples": {
                                    "NotFound": {
                                        "summary": "Nenhum vídeo encontrado",
                                        "value": { "statusCode": 404, "message": "Videos not found" }
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": ["Video"]
            }
        },
        "/video/{id}": {
            "get": {
                "summary": "Busca vídeo por ID",
                "description": "Retorna os metadados de um vídeo pelo seu ID.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do vídeo a ser consultado.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Metadados do vídeo.",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/VideoPresenter" }
                            }
                        }
                    },
                    "404": {
                        "description": "Vídeo não encontrado.",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                                "examples": {
                                    "NotFound": {
                                        "summary": "Vídeo não encontrado",
                                        "value": { "statusCode": 404, "message": "Video not found" }
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": ["Video"]
            },
            "patch": {
                "summary": "Atualiza metadados do vídeo",
                "description": "Atualiza o status e/ou chave do zip de um vídeo existente.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do vídeo a ser atualizado.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "description": "Novo status do vídeo."
                                    },
                                    "savedZipKey": {
                                        "type": "string",
                                        "description": "Chave do arquivo zip salvo.",
                                        "nullable": true
                                    }
                                },
                                "required": ["status"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Metadados do vídeo atualizados.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VideoPresenter"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição inválida.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "examples": {
                                    "MissingStatus": {
                                        "summary": "Status ausente",
                                        "value": { "statusCode": 400, "message": "Missing video status" }
                                    },
                                    "InvalidStatus": {
                                        "summary": "Status inválido",
                                        "value": { "statusCode": 400, "message": "Invalid video status value" }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vídeo não encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "examples": {
                                    "NotFound": {
                                        "summary": "Vídeo não encontrado",
                                        "value": { "statusCode": 404, "message": "Video not found" }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Video"
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter JWT Bearer token in the format 'Bearer {token}'"
            }
        },
        "schemas": {
            "VideoPresenter": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "ID do vídeo."
                    },
                    "originalVideoName": {
                        "type": "string",
                        "description": "Nome original do arquivo enviado."
                    },
                    "savedVideoKey": {
                        "type": "string",
                        "description": "Chave do arquivo salvo no armazenamento."
                    },
                    "status": {
                        "type": "string",
                        "description": "Status atual do processamento do vídeo."
                    },
                    "customerId": {
                        "type": "string",
                        "description": "ID do cliente.",
                        "nullable": true
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Data de criação.",
                        "nullable": true
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Data de atualização.",
                        "nullable": true
                    },
                    "savedZipKey": {
                        "type": "string",
                        "description": "Chave do arquivo zip salvo.",
                        "nullable": true
                    }
                },
                "required": [
                    "id",
                    "originalVideoName",
                    "savedVideoKey",
                    "status"
                ]
            },
            "AsyncUploadPresenter": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "description": "Status atual do processamento do vídeo."
                    },
                    "videoId": {
                        "type": "string",
                        "description": "ID do vídeo."
                    },
                    "message": {
                        "type": "string",
                        "description": "Mensagem de confirmação do recebimento do upload."
                    }
                },
                "required": ["status", "videoId", "message"]
            },
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "statusCode": { "type": "integer" },
                    "message": { "type": "string" }
                },
                "required": ["statusCode", "message"]
            }
        }
    }
}